refactor with base class for perm kernels
